try to remove circular dependency
