+++ src/vs/workbench/contrib/search/browser/searchActionsCopy.ts
@@ -13,9 +13,10 @@ import { KeybindingWeight } from '../../../../platform/keybinding/common/keybind
 import { KeyCode, KeyMod } from '../../../../base/common/keyCodes.js';
 import { category, getSearchView } from './searchActionsBase.js';
 import { isWindows } from '../../../../base/common/platform.js';
-import { Match } from './searchTreeModel/match.js';
-import { RenderableMatch, searchMatchComparer } from './searchTreeModel/searchTreeCommon.js';
-import { IFileInstanceMatch, IFolderMatch, IFolderMatchWithResource, isFileInstanceMatch, isFolderMatch, isFolderMatchWithResource } from './searchTreeModel/ISearchTreeBase.js';
+import { Match } from './searchTreeModel/searchTreeCommon.js';
+import { RenderableMatch } from './searchTreeModel/searchTreeCommon.js';
+import { searchMatchComparer } from './searchView.js';
+import { IFileInstanceMatch, IFolderMatch, IFolderMatchWithResource, isFileInstanceMatch, isFolderMatch, isFolderMatchWithResource } from './searchTreeModel/searchTreeCommon.js';
 
 //#region Actions
 registerAction2(class CopyMatchCommandAction extends Action2 {
