+++ src/vs/workbench/contrib/search/browser/searchTreeModel/rangeDecorations.ts
@@ -1,3 +1,8 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
 import { IDisposable, DisposableStore } from '../../../../../base/common/lifecycle';
 import { URI } from '../../../../../base/common/uri';
 import { ITextModel, TrackedRangeStickiness } from '../../../../../editor/common/model';
