+++ src/vs/workbench/contrib/search/browser/searchTreeModel/searchResult.ts
@@ -1,3 +1,8 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
 import { Event, PauseableEmitter } from '../../../../../base/common/event.js';
 import { Disposable, IDisposable } from '../../../../../base/common/lifecycle.js';
 import { URI } from '../../../../../base/common/uri.js';
@@ -8,10 +13,10 @@ import { IProgress, IProgressStep } from '../../../../../platform/progress/commo
 import { NotebookEditorWidget } from '../../../notebook/browser/notebookEditorWidget.js';
 import { INotebookEditorService } from '../../../notebook/browser/services/notebookEditorService.js';
 import { IFileMatch, ISearchComplete, ITextQuery } from '../../../../services/search/common/search.js';
-import { IFileInstanceMatch, IFolderMatch, IPlainTextSearchHeading, ISearchModel, ISearchResult, isFileInstanceMatch, isFolderMatch, isFolderMatchWithResource, isTextSearchHeading, ITextSearchHeading } from './ISearchTreeBase.js';
+import { IFileInstanceMatch, IFolderMatch, IPlainTextSearchHeading, ISearchModel, ISearchResult, isFileInstanceMatch, isFolderMatch, isFolderMatchWithResource, isTextSearchHeading, ITextSearchHeading } from './searchTreeCommon.js';
 import { AI_TEXT_SEARCH_RESULT_ID, arrayContainsElementOrParent, IChangeEvent, mergeSearchResultEvents, PLAIN_TEXT_SEARCH__RESULT_ID, RenderableMatch } from './searchTreeCommon.js';
 
-import { Match } from './match.js';
+import { Match } from './searchTreeCommon.js';
 import { RangeHighlightDecorations } from './rangeDecorations.js';
 import { PlainTextSearchHeadingImpl, TextSearchHeadingImpl } from './textSearchHeading.js';
