+++ src/vs/workbench/contrib/search/browser/searchTreeModel/searchViewModelWorkbenchService.ts
@@ -4,7 +4,7 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { createDecorator } from '../../../../../platform/instantiation/common/instantiation';
-import { ISearchModel } from './ISearchTreeBase';
+import { ISearchModel } from './searchTreeCommon';
 
 export const ISearchViewModelWorkbenchService = createDecorator<ISearchViewModelWorkbenchService>('searchViewModelWorkbenchService');
